Improve form handling for content blocks
Using custom forms for content blocks can be problematic and cause
circular import issues. It is also quite tricky to simply modify a form
for an existing content block.

I am "fixing" this limitation by making it possible to register forms
for content blocks using ``register_block_form``. When a form is
required for a content block, the previous functionality is still
working the same. However, if a form is registered for a model, it takes
precedence and is returned when the form class is looked up for this
block.

This in itself does not avoid circular import. To avoid this issue,
forms can now be specified using import paths such as
``fancypages.dashboard.forms.BlockForm`` and are only imported the first
time the form class for a model is looked up. At that point all models
should be fully loaded.