Merge standalone and oscar sandboxes
Sharing code between the standalone and Oscar sandbox is tricky with the
current setup. To reduce the amount of code duplication, the sandboxes
have been merged into the same project using django-configurations to
switch the settings. This also simplifies the settings for the tests
and uses the appropriate sandbox settings instead of re-defining them.

Resolves #51.